PEP-8 pulp.server.db.model.dispatch and its tests.
