+++ server/test/unit/server/db/model/test_dispatch.py
@@ -6,17 +6,48 @@ from uuid import uuid4
 import pickle
 import time
 import unittest
+import uuid
 
 from celery.schedules import schedule as CelerySchedule
 from mongoengine import ValidationError
+from pymongo import DESCENDING
 import bson
 import celery
 import mock
 
-from pulp.common import dateutils, constants
+from .... import base
+from pulp.common import constants, dateutils
 from pulp.server.db.model.auth import User
+from pulp.server.db.model.criteria import Criteria
 from pulp.server.db.model.dispatch import TaskStatus, ScheduledCall, ScheduleEntry
 from pulp.server.managers.factory import initialize
+
+
+SCHEDULE = {
+    u'_id': u'529f4bd93de3a31d0ec77338',
+    u'args': [u'demo1', u'puppet_distributor'],
+    u'consecutive_failures': 0,
+    u'enabled': True,
+    u'failure_threshold': 2,
+    u'first_run': u'2013-12-04T15:35:53Z',
+    u'iso_schedule': u'PT1M',
+    u'kwargs': {u'overrides': {}},
+    u'last_run_at': u'2013-12-17T00:35:53Z',
+    u'last_updated': 1387218569.811224,
+    u'principal': u"(dp0\nV_id\np1\nccopy_reg\n_reconstructor\np2\n(cbson.objectid\nObjectId\np3\n"
+                  u"c__builtin__\nobject\np4\nNtp5\nRp6\nS'R \\xab\\x06\\xe1\\x9a\\x00\\x10\\xe1i"
+                  u"\\x05\\x89'\np7\nbsVname\np8\nVadmin\np9\nsVroles\np10\n(lp11\nVsuper-users\n"
+                  u"p12\nasV_ns\np13\nVusers\np14\nsVlogin\np15\nVadmin\np16\nsVpassword\np17\n"
+                  u"VV76Yol1XYgM=,S/G6o5UyMrn0xAwbQCqFcrXnfXTh84RWhunanCDkSCo=\np18\nsVid\np19\n"
+                  u"V5220ab06e19a0010e1690589\np20\ns.",
+    u'remaining_runs': None,
+    u'resource': u'pulp:distributor:demo:puppet_distributor',
+    u'schedule': u"ccopy_reg\n_reconstructor\np0\n(ccelery.schedules\nschedule\np1\nc__builtin__\n"
+                 u"object\np2\nNtp3\nRp4\n(dp5\nS'relative'\np6\nI00\nsS'nowfun'\np7\n"
+                 u"NsS'run_every'\np8\ncdatetime\ntimedelta\np9\n(I0\nI60\nI0\ntp10\nRp11\nsb.",
+    u'task': u'pulp.server.tasks.repository.publish',
+    u'total_run_count': 1087,
+}
 
 
 initialize()
@@ -174,8 +205,9 @@ class TestTaskStatus(unittest.TestCase):
         # Invalid progress_report
         invalid_progress_reports = [4, uuid4(), object(), 'tags', [1, 2], ()]
         for invalid_progress_report in invalid_progress_reports:
-            self.assertRaises(ValidationError, TaskStatus(task_id=task_id,
-                                                          progress_report=invalid_progress_report).save)
+            self.assertRaises(
+                ValidationError,
+                TaskStatus(task_id=task_id, progress_report=invalid_progress_report).save)
 
     def test_task_type_validation(self):
         # Valid task_type
@@ -940,28 +972,225 @@ class TestScheduleEntryIsDue(unittest.TestCase):
         self.assertEqual(seconds, 1442)
 
 
-SCHEDULE = {
-    u'_id': u'529f4bd93de3a31d0ec77338',
-    u'args': [u'demo1', u'puppet_distributor'],
-    u'consecutive_failures': 0,
-    u'enabled': True,
-    u'failure_threshold': 2,
-    u'first_run': u'2013-12-04T15:35:53Z',
-    u'iso_schedule': u'PT1M',
-    u'kwargs': {u'overrides': {}},
-    u'last_run_at': u'2013-12-17T00:35:53Z',
-    u'last_updated': 1387218569.811224,
-    u'principal': u"(dp0\nV_id\np1\nccopy_reg\n_reconstructor\np2\n(cbson.objectid\nObjectId\np3\n"
-                  u"c__builtin__\nobject\np4\nNtp5\nRp6\nS'R \\xab\\x06\\xe1\\x9a\\x00\\x10\\xe1i"
-                  u"\\x05\\x89'\np7\nbsVname\np8\nVadmin\np9\nsVroles\np10\n(lp11\nVsuper-users\n"
-                  u"p12\nasV_ns\np13\nVusers\np14\nsVlogin\np15\nVadmin\np16\nsVpassword\np17\n"
-                  u"VV76Yol1XYgM=,S/G6o5UyMrn0xAwbQCqFcrXnfXTh84RWhunanCDkSCo=\np18\nsVid\np19\n"
-                  u"V5220ab06e19a0010e1690589\np20\ns.",
-    u'remaining_runs': None,
-    u'resource': u'pulp:distributor:demo:puppet_distributor',
-    u'schedule': u"ccopy_reg\n_reconstructor\np0\n(ccelery.schedules\nschedule\np1\nc__builtin__\n"
-                 u"object\np2\nNtp3\nRp4\n(dp5\nS'relative'\np6\nI00\nsS'nowfun'\np7\n"
-                 u"NsS'run_every'\np8\ncdatetime\ntimedelta\np9\n(I0\nI60\nI0\ntp10\nRp11\nsb.",
-    u'task': u'pulp.server.tasks.repository.publish',
-    u'total_run_count': 1087,
-}
+class TaskStatusTests(base.PulpServerTests):
+    """
+    Test the TaskStatus class functions.
+    """
+    def clean(self):
+        super(TaskStatusTests, self).clean()
+        TaskStatus.objects().delete()
+
+    def get_random_uuid(self):
+        return str(uuid.uuid4())
+
+    def test_create_task_status(self):
+        """
+        Tests that TaskStatus creation with valid data is successful.
+        """
+        task_id = self.get_random_uuid()
+        worker_name = 'a_worker_name'
+        tags = ['test-tag1', 'test-tag2']
+        state = 'waiting'
+
+        created = TaskStatus(task_id, worker_name, tags, state).save()
+
+        task_statuses = TaskStatus.objects()
+        self.assertEqual(1, len(task_statuses))
+
+        task_status = task_statuses[0]
+        self.assertEqual(task_id, task_status['task_id'])
+        self.assertEqual(worker_name, task_status['worker_name'])
+        self.assertEqual(tags, task_status['tags'])
+        self.assertEqual(state, task_status['state'])
+
+        self.assertEqual(task_id, created['task_id'])
+        self.assertEqual(tags, created['tags'])
+        self.assertEqual(state, created['state'])
+
+    def test_create_task_status_defaults(self):
+        """
+        Tests TaskStatus creation with minimal information, to ensure that defaults are handled
+        properly.
+        """
+        task_id = self.get_random_uuid()
+
+        TaskStatus(task_id).save()
+
+        task_statuses = TaskStatus.objects()
+        self.assertEqual(1, len(task_statuses))
+        self.assertEqual(task_id, task_statuses[0]['task_id'])
+        self.assertEqual(None, task_statuses[0]['worker_name'])
+        self.assertEqual([], task_statuses[0]['tags'])
+        self.assertEqual('waiting', task_statuses[0]['state'])
+
+    def test_create_task_status_invalid_task_id(self):
+        """
+        Test that TaskStatus creation with an invalid task id raises the correct error.
+        """
+        try:
+            TaskStatus(None).save()
+        except ValidationError, e:
+            self.assertTrue('task_id' in e.message)
+        else:
+            self.fail('Invalid ID did not raise an exception')
+
+    def test_create_task_status_duplicate_task_id(self):
+        """
+        Tests TaskStatus creation with a duplicate task id.
+        """
+        task_id = self.get_random_uuid()
+
+        TaskStatus(task_id).save()
+        try:
+            TaskStatus(task_id).save()
+        except Exception, e:
+            self.assertTrue(task_id in e.message)
+        else:
+            self.fail('Task status with a duplicate task id did not raise an exception')
+
+    def test_create_task_status_invalid_attributes(self):
+        """
+        Tests that TaskStatus creation with invalid attributes
+        results in an error
+        """
+        task_id = self.get_random_uuid()
+        worker_name = ['not a string']
+        tags = 'not a list'
+        state = 1
+        try:
+            TaskStatus(task_id, worker_name, tags, state).save()
+        except ValidationError, e:
+            self.assertTrue('tags' in e.message)
+            self.assertTrue('state' in e.message)
+            self.assertTrue('worker_name' in e.message)
+        else:
+            self.fail('Invalid attributes did not cause create to raise an exception')
+
+    def test_task_status_update(self):
+        """
+        Tests the successful operation of task status update.
+        """
+        task_id = self.get_random_uuid()
+        worker_name = 'special_worker_name'
+        tags = ['test-tag1', 'test-tag2']
+        state = 'waiting'
+        TaskStatus(task_id, worker_name, tags, state).save()
+        now = datetime.now(dateutils.utc_tz())
+        start_time = dateutils.format_iso8601_datetime(now)
+        delta = {'start_time': start_time,
+                 'state': 'running',
+                 'progress_report': {'report-id': 'my-progress'}}
+
+        TaskStatus.objects(task_id=task_id).update_one(
+            set__start_time=delta['start_time'], set__state=delta['state'],
+            set__progress_report=delta['progress_report'])
+
+        task_status = TaskStatus.objects(task_id=task_id).first()
+        self.assertEqual(task_status['start_time'], delta['start_time'])
+        # Make sure that parse_iso8601_datetime is able to parse the start_time without errors
+        dateutils.parse_iso8601_datetime(task_status['start_time'])
+        self.assertEqual(task_status['state'], delta['state'])
+        self.assertEqual(task_status['progress_report'], delta['progress_report'])
+        self.assertEqual(task_status['worker_name'], worker_name)
+
+    @mock.patch('pulp.server.db.model.base.CriteriaQuerySet.find_by_criteria')
+    def test_find_by_criteria(self, mock_find_by_criteria):
+        criteria = Criteria()
+        TaskStatus.objects.find_by_criteria(criteria)
+        mock_find_by_criteria.assert_called_once_with(criteria)
+
+    def test_find_by_criteria_with_result(self):
+        tags = ['test', 'tags']
+        TaskStatus(task_id='1').save()
+        TaskStatus(task_id='2', tags=tags).save()
+
+        result = 'done'
+        TaskStatus(task_id='3', tags=tags, state=constants.CALL_FINISHED_STATE,
+                   result=result).save()
+
+        filters = {'tags': tags, 'task_id': {'$in': ['1', '3']}}
+        fields = ['task_id', 'tags', 'result']
+        limit = 1
+        sort = (('task_id', DESCENDING), )
+        criteria = Criteria(filters=filters, fields=fields, limit=limit, sort=sort)
+        query_set = TaskStatus.objects.find_by_criteria(criteria)
+        self.assertEqual(len(query_set), 1)
+        self.assertEqual(query_set[0].task_id, '3')
+        self.assertEqual(query_set[0].result, result)
+        task_state_default = constants.CALL_WAITING_STATE
+        self.assertEqual(query_set[0].state, task_state_default)
+
+    def test_set_accepted(self):
+        task_id = self.get_random_uuid()
+        TaskStatus(task_id, state=constants.CALL_WAITING_STATE).save()
+
+        TaskStatus.objects(task_id=task_id, state=constants.CALL_WAITING_STATE).\
+            update_one(set__state=constants.CALL_ACCEPTED_STATE)
+        task_status = TaskStatus.objects.get(task_id=task_id)
+        self.assertTrue(task_status['state'], constants.CALL_ACCEPTED_STATE)
+
+    @mock.patch('pulp.common.dateutils.format_iso8601_datetime')
+    def test_set_succeeded(self, mock_date):
+        task_id = self.get_random_uuid()
+        TaskStatus(task_id).save()
+
+        result = 'done'
+        now = '2014-11-21 05:21:38.829678'
+        mock_date.return_value = now
+
+        t = datetime.now(dateutils.utc_tz())
+        finished = dateutils.format_iso8601_datetime(t)
+        TaskStatus.objects(task_id=task_id).update_one(set__finish_time=finished,
+                                                       set__state=constants.CALL_FINISHED_STATE,
+                                                       set__result=result)
+        task_status = TaskStatus.objects(task_id=task_id).first()
+        self.assertTrue(task_status['state'], constants.CALL_FINISHED_STATE)
+        self.assertTrue(task_status['finish_time'], now)
+        self.assertTrue(task_status['result'], result)
+
+    def test_set_succeeded_with_timestamp(self):
+        task_id = self.get_random_uuid()
+        TaskStatus(task_id).save()
+
+        result = 'done'
+        now = '2014-11-21 05:21:38.829678'
+
+        TaskStatus.objects(task_id=task_id).update_one(set__finish_time=now,
+                                                       set__state=constants.CALL_FINISHED_STATE,
+                                                       set__result=result)
+        task_status = TaskStatus.objects(task_id=task_id).first()
+        self.assertTrue(task_status['state'], constants.CALL_FINISHED_STATE)
+        self.assertTrue(task_status['finish_time'], now)
+        self.assertTrue(task_status['result'], result)
+
+    @mock.patch('pulp.common.dateutils.format_iso8601_datetime')
+    def test_set_failed(self, mock_date):
+        task_id = self.get_random_uuid()
+        TaskStatus(task_id).save()
+
+        traceback = 'abcdef'
+        finished = '2014-11-21 05:21:38.829678'
+        mock_date.return_value = finished
+
+        TaskStatus.objects(task_id=task_id).update_one(set__finish_time=finished,
+                                                       set__state=constants.CALL_ERROR_STATE,
+                                                       set__traceback=traceback)
+        task_status = TaskStatus.objects.get(task_id=task_id)
+        self.assertTrue(task_status['state'], constants.CALL_ERROR_STATE)
+        self.assertTrue(task_status['finish_time'], finished)
+        self.assertTrue(task_status['traceback'], traceback)
+
+    def test_set_failed_with_timestamp(self):
+        task_id = self.get_random_uuid()
+        TaskStatus(task_id).save()
+
+        traceback = 'abcdef'
+        finished = '2014-11-21 05:21:38.829678'
+
+        TaskStatus.objects(task_id=task_id).update_one(set__finish_time=finished,
+                                                       set__state=constants.CALL_ERROR_STATE,
+                                                       set__traceback=traceback)
+        task_status = TaskStatus.objects.get(task_id=task_id)
+        self.assertTrue(task_status['state'], constants.CALL_ERROR_STATE)
+        self.assertTrue(task_status['finish_time'], finished)
+        self.assertTrue(task_status['traceback'], traceback)
